Moves NotificationItem in a seperate file
Auditors: @bsclifton

Test Plan:
